refactor: ⚡️ 次のターンの生成処理を関数として切り出し
