+++ src/game/logics/boardLogic.ts
@@ -52,66 +52,3 @@ export const updateTileStatus = (
 
   return newBoard;
 };
-
-/**
- * ゲームが終了したかを判定する
- */
-export const checkForWin = (currentTurn: Turn, board: Board): boolean => {
-  // 各タイルに対して、位置情報と状態をオブジェクトとしてマッピング
-  const tileStates = board.tiles.map((tile) => ({
-    tile: tile,
-    state: getTileState(currentTurn, board.size, tile),
-  }));
-
-  // 横の行をチェック
-  for (let y = 0; y < board.size; y++) {
-    const row = tileStates.filter(({ tile }) => tile.y === y);
-    if (
-      row.every(
-        ({ state }) =>
-          state.symbol !== "" && state.symbol === row[0].state.symbol,
-      )
-    ) {
-      return true;
-    }
-  }
-
-  // 縦の列をチェック
-  for (let x = 0; x < board.size; x++) {
-    const column = tileStates.filter(({ tile }) => tile.x === x);
-    if (
-      column.every(
-        ({ state }) =>
-          state.symbol !== "" && state.symbol === column[0].state.symbol,
-      )
-    ) {
-      return true;
-    }
-  }
-
-  // 斜めのチェック (左上から右下)
-  const diagonal1 = tileStates.filter(({ tile }) => tile.x === tile.y);
-  if (
-    diagonal1.every(
-      ({ state }) =>
-        state.symbol !== "" && state.symbol === diagonal1[0].state.symbol,
-    )
-  ) {
-    return true;
-  }
-
-  // 斜めのチェック (右上から左下)
-  const diagonal2 = tileStates.filter(
-    ({ tile }) => tile.x === board.size - 1 - tile.y,
-  );
-  if (
-    diagonal2.every(
-      ({ state }) =>
-        state.symbol !== "" && state.symbol === diagonal2[0].state.symbol,
-    )
-  ) {
-    return true;
-  }
-
-  return false;
-};
